requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
apscheduler>=3.10.0
requests>=2.32.0
sentence-transformers>=3.0.0
transformers>=4.41.0
torch>=2.6.0
bertopic>=0.16.0
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0
streamlit>=1.35.0
plotly>=5.22.0
httpx>=0.27.0
pydantic>=2.7.0
umap-learn>=0.5.6
hdbscan>=0.8.38